refactor: ajuste de pastas e retirando warnings
