+++ src/main/java/com/thalos/trailerflix/dtos/exception/FieldMessageDTO.java
@@ -1,4 +1,4 @@
-package com.thalos.trailerflix.exceptions;
+package com.thalos.trailerflix.dtos.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -11,9 +11,9 @@ import java.io.Serializable;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class FieldMessage implements Serializable {
-    private static final long serialVersionUid = 1L;
-
-    private String fieldName;
+public class FieldMessageDTO implements Serializable {
+	private static final long serialVersionUID = 1L;
+	
+	private String fieldName;
     private String message;
 }
