+++ src/main/java/com/thalos/trailerflix/exceptions/handler/ResourceExceptionHandler.java
@@ -1,7 +1,10 @@
-package com.thalos.trailerflix.exceptions;
+package com.thalos.trailerflix.exceptions.handler;
 
-import com.thalos.trailerflix.exceptions.exception.InternalServerException;
-import com.thalos.trailerflix.exceptions.exception.ObjectNotFoundException;
+import com.thalos.trailerflix.dtos.exception.StandardErrorDTO;
+import com.thalos.trailerflix.dtos.exception.ValidationErrorDTO;
+import com.thalos.trailerflix.exceptions.InternalServerException;
+import com.thalos.trailerflix.exceptions.ObjectNotFoundException;
+
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,30 +16,30 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 @ControllerAdvice
 public class ResourceExceptionHandler {
     @ExceptionHandler(ObjectNotFoundException.class)
-    public ResponseEntity<StandardError> objectNotFoundException(ObjectNotFoundException e) {
-        StandardError standardError = new StandardError(System.currentTimeMillis(),
+    public ResponseEntity<StandardErrorDTO> objectNotFoundException(ObjectNotFoundException e) {
+        StandardErrorDTO standardError = new StandardErrorDTO(System.currentTimeMillis(),
                 HttpStatus.NOT_FOUND.value(), e.getMessage());
 
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(standardError);
     }
 
     @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<StandardError> dataIntegrityViolationException(DataIntegrityViolationException e) {
-        StandardError standardError = new StandardError(System.currentTimeMillis(),
+    public ResponseEntity<StandardErrorDTO> dataIntegrityViolationException(DataIntegrityViolationException e) {
+        StandardErrorDTO standardError = new StandardErrorDTO(System.currentTimeMillis(),
                 HttpStatus.BAD_REQUEST.value(), e.getMessage());
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(standardError);
     }
 
     @ExceptionHandler(InternalServerException.class)
-    public ResponseEntity<StandardError> internalServerException(InternalServerException e) {
-        StandardError standardError = new StandardError(System.currentTimeMillis(),
+    public ResponseEntity<StandardErrorDTO> internalServerException(InternalServerException e) {
+        StandardErrorDTO standardError = new StandardErrorDTO(System.currentTimeMillis(),
                 HttpStatus.INTERNAL_SERVER_ERROR.value(), e.getMessage());
         return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(standardError);
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<StandardError> methodArgumentNotValidException(MethodArgumentNotValidException e) {
-        ValidationError validationError = new ValidationError(System.currentTimeMillis(),
+    public ResponseEntity<StandardErrorDTO> methodArgumentNotValidException(MethodArgumentNotValidException e) {
+        ValidationErrorDTO validationError = new ValidationErrorDTO(System.currentTimeMillis(),
                 HttpStatus.BAD_REQUEST.value(), "Erro na validação dos campos");
 
         for (FieldError fieldMessage : e.getBindingResult().getFieldErrors()) {
